Move Version to components/common
