+++ airbyte-webapp/src/components/Indicator/index.tsx
@@ -1,4 +1,4 @@
-import Indicator from "./Indicator";
+import { Indicator } from "./Indicator";
 
 export default Indicator;
 export { Indicator };
