--- airbyte-webapp/src/components/Version/index.tsx
@@ -1,3 +1,0 @@
-import Version from "./Version";
-
-export default Version;
