+++ airbyte-webapp/src/components/common/Version/index.ts
@@ -1,0 +1,1 @@
+export * from "./Version";
